requirements.txt
pandas==1.3.5
requests==2.26.0
google-cloud-bigquery==2.34.3
pyyaml==6.0
apache-airflow==2.7.0
pytest==7.1.2
mysql-connector-python==8.0.29